fergo-server/requirements.txt
django
djangorestframework
djangorestframework-jsonapi